Add PDF with minimal knowledge expected from all directors.
Thank you for attending the NYS Attorney General’s 2024 Plattsburgh Charities Symposium!
As promised, we are forwarding the materials we distributed at the event,
as well as copies of the slide deck presentations from James Sheehan and
Kelly Mathews together with materials referenced during their presentations.
Due to the size of the materials, we will send them in two circulations,
so please expect a second communication shortly after this.
We hope this information is helpful to you, and we look forward to seeing you
next time!
Have a great summer!
Sincerely,
The NYS Attorney General’s Charities Bureau and Plattsburgh Regional Office